requirements.txt
asgiref==3.8.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
flake8==6.1.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.6
isodate==0.6.1
mccabe==0.7.0
pillow==10.3.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
pytils==0.4.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.1.0
drf-yasg==1.21.7
currency.converter==0.5.5
stripe~=9.6.0
celery==5.4.0